Many, many big changes. 	1. Change many floating-point types from f32 to f64 (more consistency overall that way). 	2. Move all repeated time code to a new file, time.rs. Less repetition with `chrono` that way. 	3. Allow for textures to have transition animations, that vary the aspect ratio and opacity in different ways. These are controlled by custom easing functions. Also allow for explicit control over the scroll speed, without changing the easing function. 	4. Make a library that includes a bunch of helpful easing functions, so that they don't need to be repeated across lots of different files, and so that they can be made more complex without diverting too much from the core logic of each such file. 	5. When rendering windows in the window tree, localize errors from window updaters, so that one error doesn't stop other windows from updating/rendering. Do the same for drawing window contents. 	6. in continually_updated.rs, restruture much of the code there. Use native Tokio channels, and use them in an asynchronous manner, which should allow `ContinuallyUpdated` tasks to wait for data / send data without blocking Tokio's scheduler. Overall, this should mean that tasks depending on `ContinuallyUpdated` should be able to complete quicker.
Cargo.toml: upgrade sdl2 and chrono. add a note about replacing `lazy_static` at some point with `once_cell`.

dashboard_defs:
	clock.rs: use f64, and the new time system.
	credit.rs: use a named scrolling easing function, and specify a scroll speed multiplier.
	error.rs: reorder 2 imports, use a named scrolling easing function, and specify a scroll speed multiplier.
	mod.rs: add the new easing functions to this.
	spinitron.rs: use a named scrolling easing function, and specify a scroll speed multiplier.
	surprise.rs: use the new time system. Remove a cast.
	themes/*.rs: use the new time system. Remove a `pub` qualifier.
	twilio.rs: clean up some imports. Use the new time system, and the new easing fn library. Remove an old TODO. Use f64 over f32.
	updatable_text_pattern.rs: specify a scroll easer (along with a scroll speed multiplier) in the updatable text pattern code.
	weather.rs: use a named scrolling easing function, and specify a scroll speed multiplier. Use an `await` in a new place for the new `ContinuallyUpdated` implementation.
	easing_fns.rs: see big change #4.

spinitron:
	api.rs: remove a newline.
	model.rs: add some TODOs. Use the new time system. Fix a rare bug for show timestamps.
	state.rs: use the new time system, and the new continually updated change as well.

utility_types:
	continually_updated.rs: see big change #6.
	mod.rs: add time.rs.
	vec2f: make the component type to f64 (it was f32 before).
	time.rs: see big change #2.

src/
	main.rs: specify a max queue size for remake transitions, and do various other formatting tweaks as well. Also handle the change to how errors are handled across different windows as well.
	window_tree.rs: make a `PreciseRect` type, to avoid more imprecision issues (`FRect` exists in this newest version of SDL2 now). Use SDL2's `FPoint`, instead of `Point`, and `FRect` when needed. See big change #5 as well. Handle the aspect ratio changes relevant to texture transition animations too.
	texture.rs: implement texture transitions. This works for text->plain textures, and vice versa. So much testing has been done for this. Setting blend modes / alpha mods still needs to be implemented for this. Behind the scenes, textures are queued up, and both are displayed using interpolation of aspect ratios and opacities.

+++ src/utility_types/vec2f.rs
@@ -2,7 +2,7 @@ use std::ops;
 
 /* TODO: could I wrap this in a struct type?
 That would then allow for things like closure under multiplication */
-type Component = f32;
+type Component = f64;
 
 // A 0-1 normalized floating-point vec2
 #[derive(Copy, Clone, PartialEq)]
